fix sonar and changed testfile executing
