+++ sonar-project.properties
@@ -2,9 +2,9 @@ sonar.organization=ita-social-projects
 sonar.projectKey=ita-social-projects-horondi-client-admin
 sonar.sources=. 
 sonar.host.url=https://sonarcloud.io
-sonar.exclusions=public, src/**/*.test.js, src/**/**/*.test.js, src/**/**/**/*.test.js, src/**/**/**/**/*.test.js, src/**/**/**/**/**/*.test.js
+sonar.exclusions=public, **/*.spec.js, **/*.test.js, cypress/**, src/locales/**/index.js, **/__mocks__/**, src/setupTests.js
 sonar.language=js
 sonar.tests=src
-sonar.test.inclusions=src/**/*.test.js, src/**/**/*.test.js, src/**/**/**/*.test.js, src/**/**/**/**/*.test.js, src/**/**/**/**/**/*.test.js, cypress/**
+sonar.test.inclusions=**/*.spec.js, **/*.test.js
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.testExecutionReportPaths=test-report.xml
