+++ src/components/forms/basics-form/tests/basics-form.spec.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import * as reactRedux from 'react-redux';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import { TextField } from '@material-ui/core';
-
-import UserForm from '../index';
+import { Paper, Grid } from '@material-ui/core';
+import LoadingBar from '../../../loading-bar';
+import BasicsForm from '../index';
 import ImageUploadContainer from '../../../../containers/image-upload-container';
+import MaterialsContainer from '../../../../containers/materials-container';
+import CheckboxOptions from '../../../checkbox-options';
 import { config } from '../../../../configs';
-import { files, target } from './user.form.variables';
-import { SaveButton } from '../../../buttons';
+import { mockMaterial, files, target } from './basics-form.variables';
 import FileReaderMock from '../../../../../__mocks__/fileReaderMock';
 
 const mockSetFieldValue = jest.fn();
@@ -16,7 +15,7 @@ const mockSubmit = jest.fn();
 const mockChange = jest.fn();
 const mockBlur = jest.fn();
 const mockSetUpload = jest.fn();
-const mockSetUserImage = jest.fn();
+const mockSetBasicImage = jest.fn();
 
 const { GO_BACK_TITLE, SAVE_TITLE } = config.buttonTitles;
 
@@ -27,59 +26,42 @@ jest.mock('formik', () => ({
     handleSubmit: mockSubmit,
     handleChange: mockChange,
     touched: {
-      userFirstName: 'a',
-      userLastName: 'a',
-      email: 'aaaa',
-      phoneNumber: '3650',
-      country: '12',
-      region: '12',
-      city: '12',
-      street: '12',
-      buildingNumber: 'aaaa',
-      appartment: 'aaaa',
-      zipcode: '12'
+      basicImage: 'image',
+      additionalPrice: 1
     },
     errors: {
-      userFirstName: 'a',
-      userLastName: 'a',
-      email: 'aaaa',
-      phoneNumber: '3650',
-      country: '12',
-      region: '12',
-      city: '12',
-      street: '12',
-      buildingNumber: 'aaaa',
-      appartment: 'aaaa',
-      zipcode: '12'
+      basicImage: 'image',
+      additionalPrice: 1
     },
     setFieldValue: mockSetFieldValue,
     handleBlur: mockBlur
   })
 }));
 jest.mock('../../../../hooks/form-dialog/use-unsaved-changes-handler');
-jest.mock('../../../../utils/use-user-handlers', () => ({
+jest.mock('../../../../utils/use-basics-handlers', () => ({
   __esModule: true,
   default: () => ({
     setUpload: mockSetUpload,
-    setUserImage: mockSetUserImage
+    setBasicImage: mockSetBasicImage
   })
 }));
 
 const fileReader = new FileReaderMock();
 jest.spyOn(window, 'FileReader').mockImplementation(() => fileReader);
 
-describe('User form tests', () => {
+describe('Basics form tests', () => {
   let spyOnUseSelector;
   let spyOnUseDispatch;
   let component;
 
   beforeEach(() => {
     spyOnUseSelector = jest.spyOn(reactRedux, 'useSelector');
+    spyOnUseSelector.mockImplementation(() => mockMaterial);
 
     spyOnUseDispatch = jest.spyOn(reactRedux, 'useDispatch');
 
     spyOnUseDispatch.mockImplementation(() => jest.fn());
-    component = mount(<UserForm />);
+    component = mount(<BasicsForm />);
   });
   afterEach(() => {
     component.unmount();
@@ -88,7 +70,7 @@ describe('User form tests', () => {
   });
 
   it('should render form component', () => {
-    expect(component.find('form').length).toBe(1);
+    expect(component.find('form').length).toBe(2);
   });
 
   it('should call preventDefault method', () => {
@@ -100,20 +82,42 @@ describe('User form tests', () => {
     expect(event.preventDefault).toHaveBeenCalled();
   });
 
-  it('Should render 11 components of TextField type', () => {
-    expect(component.find(TextField)).toHaveLength(11);
+  it('should render CheckboxOptions component', () => {
+    const wrapper = component.find(CheckboxOptions);
+    expect(wrapper).toHaveLength(1);
   });
 
-  it('should render ImageUploadContainer component', () => {
-    const wrapper = component.find(ImageUploadContainer);
+  it('should render MaterialsContainer component', () => {
+    const wrapper = component.find(MaterialsContainer);
     expect(wrapper).toHaveLength(1);
+  });
+
+  it('Should render 2 buttons and 6 inputs', () => {
+    expect(component.find('input')).toHaveLength(6);
+    expect(component.find('button')).toHaveLength(2);
   });
 
   it(`Should render go-back button with '${GO_BACK_TITLE}' label`, () => {
     expect(component.find('button').at(0).text()).toBe(GO_BACK_TITLE);
   });
+
   it(`Should render save button with '${SAVE_TITLE}' label`, () => {
     expect(component.find('button').at(1).text()).toBe(SAVE_TITLE);
+  });
+
+  it('should render Grid component', () => {
+    const wrapper = component.find(Grid);
+    expect(wrapper.exists(Grid)).toBeDefined();
+  });
+
+  it('should render Paper component', () => {
+    const wrapper = component.find(Paper);
+    expect(wrapper.exists(Paper)).toBeDefined();
+  });
+
+  it('should render ImageUploadPreviewContainer component', () => {
+    const wrapper = component.find(ImageUploadContainer);
+    expect(wrapper.exists(ImageUploadContainer)).toBeDefined();
   });
 
   it('Should upload image', () => {
@@ -124,7 +128,7 @@ describe('User form tests', () => {
     expect(mockSetUpload).toHaveBeenCalledWith(files[0]);
   });
 
-  it('Should test FileReader ', () => {
+  it('Should test FileReader ', async () => {
     fileReader.result = 'file content';
     const imageContainer = component.find(ImageUploadContainer);
     const handler = imageContainer.prop('handler');
@@ -134,8 +138,16 @@ describe('User form tests', () => {
     expect(fileReader.readAsDataURL).toHaveBeenCalled();
     expect(fileReader.readAsDataURL).toHaveBeenCalledWith(files[0]);
   });
-  it('Should simulate submit button', () => {
-    component.find(SaveButton).prop('onClickHandler')();
-    expect(mockSubmit).toHaveBeenCalled();
+
+  it('should call setFieldValue for checkbox', () => {
+    component.find('CheckboxOptions').props().options[0].handler();
+    expect(mockSetFieldValue).toHaveBeenCalledWith('available', true);
+  });
+
+  it('Loading bar should be visible', () => {
+    mockMaterial.loading = true;
+    component = mount(<BasicsForm />);
+    const loadingBar = component.find(LoadingBar);
+    expect(loadingBar).toBeDefined();
   });
 });
