+++ src/components/forms/contacts-form/tests/contacts-form.spec.js
@@ -1,7 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import * as reactRedux from 'react-redux';
 import ContactsForm from '../index';
 import { config } from '../../../../configs';
