+++ src/components/forms/tests/back/backs-form.spec.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import * as reactRedux from 'react-redux';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import PropTypes from 'prop-types';
 import FormControl from '@material-ui/core/FormControl';
 import InputLabel from '@material-ui/core/InputLabel';
@@ -11,7 +9,7 @@ import { Paper, Grid } from '@material-ui/core';
 import LanguagePanel from '../../language-panel';
 import LoadingBar from '../../../loading-bar';
 import { BackButton, SaveButton } from '../../../buttons';
-import BottomForm from '../../bottom-form/index';
+import BackForm from '../../back-form/index';
 import ImageUploadPreviewContainer from '../../../../containers/image-upload-container/image-upload-previewContainer';
 import CheckboxOptions from '../../../checkbox-options';
 import { config } from '../../../../configs';
@@ -27,11 +25,10 @@ const mockSubmit = jest.fn();
 const mockChange = jest.fn();
 const mockBlur = jest.fn();
 const mockSetUpload = jest.fn();
-const mockSetBottomImage = jest.fn();
+const mockSetBackImage = jest.fn();
 
 const { GO_BACK_TITLE, SAVE_TITLE } = config.buttonTitles;
 
-jest.mock('../../../../hooks/form-dialog/use-unsaved-changes-handler');
 jest.mock('formik', () => ({
   ...jest.requireActual('formik'),
   useFormik: () => ({
@@ -45,23 +42,29 @@ jest.mock('formik', () => ({
   })
 }));
 
-jest.mock('../../../../utils/use-bottom-handlers.js', () => ({
+jest.mock('../../../../utils/use-back-handlers.js', () => ({
   __esModule: true,
   default: () => ({
     setUpload: mockSetUpload,
-    setBottomImage: mockSetBottomImage
+    setBackImage: mockSetBackImage
   })
 }));
+jest.mock('../../../../hooks/form-dialog/use-unsaved-changes-handler');
 
-const fileReader = new FileReaderMock();
-jest.spyOn(window, 'FileReader').mockImplementation(() => fileReader);
+jest.spyOn(global, 'FileReader').mockImplementation(function () {
+  this.readAsDataURL = jest.fn();
+  this.onload = jest.fn();
+});
 
-describe('Bottom form tests', () => {
+describe('Back form tests', () => {
   let spyOnUseSelector;
   let spyOnUseDispatch;
   let mockDispatch;
   let component;
   let getState;
+
+  const fileReader = new FileReaderMock();
+  jest.spyOn(window, 'FileReader').mockImplementation(() => fileReader);
 
   beforeEach(() => {
     spyOnUseSelector = jest.spyOn(reactRedux, 'useSelector');
@@ -71,7 +74,7 @@ describe('Bottom form tests', () => {
     mockDispatch = jest.fn();
 
     spyOnUseDispatch.mockImplementation(() => jest.fn());
-    component = mount(<BottomForm />);
+    component = mount(<BackForm />);
   });
   afterEach(() => {
     component.unmount();
@@ -90,12 +93,12 @@ describe('Bottom form tests', () => {
     expect(CheckboxOptions).toHaveLength(1);
   });
 
-  it('Should render 2 buttons and 5 inputs', () => {
+  it('Should render 2 buttons and 6 inputs', () => {
     expect(component.find('input')).toHaveLength(6);
     expect(component.find('button')).toHaveLength(2);
   });
 
-  it(`Should render go-bottom button with '${GO_BACK_TITLE}' label`, () => {
+  it(`Should render go-back button with '${GO_BACK_TITLE}' label`, () => {
     expect(component.find('button').at(0).text()).toBe(GO_BACK_TITLE);
   });
 
@@ -134,11 +137,11 @@ describe('Bottom form tests', () => {
     const imageContainer = component.find(ImageUploadPreviewContainer);
     const handler = imageContainer.prop('handler');
     handler(files);
-
+    fileReader.onload(target);
     expect(fileReader.readAsDataURL).toHaveBeenCalled();
     expect(fileReader.readAsDataURL).toHaveBeenCalledWith(files[0]);
+    expect(mockSetBackImage).toHaveBeenCalled();
   });
-
   it('should render FormControl component', () => {
     const wrapper = component.find(FormControl);
     expect(wrapper.exists(FormControl)).toBeDefined();
@@ -183,7 +186,7 @@ describe('Bottom form tests', () => {
     expect(SaveButton).toHaveLength(1);
   });
 
-  it('Should render bottom-form', () => {
+  it('Should render back-form', () => {
     expect(component).toBeDefined();
     expect(component).toHaveLength(1);
   });
@@ -195,8 +198,8 @@ describe('Bottom form tests', () => {
   });
 
   it('Should have appropriate prop types', () => {
-    expect(BottomForm.propTypes.id).toBe(PropTypes.string);
-    expect(BottomForm.propTypes.edit).toBe(PropTypes.bool);
+    expect(BackForm.propTypes.id).toBe(PropTypes.string);
+    expect(BackForm.propTypes.edit).toBe(PropTypes.bool);
   });
 
   it('Loading bar should be not visible', () => {
@@ -205,7 +208,7 @@ describe('Bottom form tests', () => {
 
   it('Loading bar should be visible', () => {
     mockMaterial.loading = true;
-    component = mount(<BottomForm />);
+    component = mount(<BackForm />);
     const loadingBar = component.find(LoadingBar);
     expect(component.exists(LoadingBar)).toBeDefined();
     expect(component.exists(LoadingBar)).toBe(true);
