+++ src/pages/comments/comment-edit/replyComments/tests/reply-comments.spec.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import * as redux from 'react-redux';
 import { config } from '../../../../../configs';
 import LoadingBar from '../../../../../components/loading-bar';
@@ -10,8 +8,6 @@ import ReplyComments from '../index';
 import { itemsCount, replyComments } from './reply-comments.variables';
 
 const { NO_REPLY_COMMENTS_MESSAGE } = config.messages;
-
-configure({ adapter: new Adapter() });
 
 const mockUseDispatchFn = jest.fn();
 const mockUseSuccessSnackbar = jest.fn();
