+++ src/pages/main-page/tests/main-page.spec.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { configure, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import {
   useSelector as useSelectorMock,
   useDispatch as useDispatchMock
@@ -22,8 +20,6 @@ jest.mock('react-redux', () => ({
   useSelector: jest.fn().mockImplementationOnce((selector) => selector()),
   useDispatch: jest.fn()
 }));
-
-configure({ adapter: new Adapter() });
 
 describe('Main-page component tests', () => {
   const mockHolder = jest.fn();
